+++ ntex-rt/CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## [0.4.16] - 2024-09-02
+
+* Allow to override block_on function
+
+* Add stack size configuration
+
 ## [0.4.15] - 2024-08-30
 
 * No runtime compatibility